Separate ExecutionPlan op construction logic into different files
